Refactor and Optimization
