+++ .idea/dictionaries/VINOTH.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="VINOTH">
+    <words>
+      <w>robo</w>
+    </words>
+  </dictionary>
+</component>
