+++ app/src/androidTest/java/com/hcl/canada/ui/AboutCanadaListActivityTest.java
@@ -4,7 +4,9 @@ import android.content.Intent;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 
-import com.hcl.canada.ui.api.APIClient;
+import com.hcl.canada.robots.AboutListRobot;
+import com.hcl.canada.api.APIClient;
+import com.hcl.canada.utils.FileUtils;
 
 import org.junit.After;
 import org.junit.Before;
@@ -14,8 +16,6 @@ import org.junit.runner.RunWith;
 
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
-import robots.AboutListRobot;
-import utils.FileUtils;
 
 import static java.net.HttpURLConnection.HTTP_OK;
 import static java.net.HttpURLConnection.HTTP_UNAVAILABLE;
@@ -26,22 +26,25 @@ public class AboutCanadaListActivityTest {
     @Rule
     public ActivityTestRule<AboutCanadaListActivity> activityTestRule = new ActivityTestRule<>(AboutCanadaListActivity.class, false, false);
 
-    public AboutListRobot.Eyes eyes = new AboutListRobot.Eyes();
+    private static final String URL_PATH = "/";
+    private static final String ABOUT_FACTS_CONTENT_JSON = "/about/facts.json";
+    private static final String ABOUT_FACTS_ERROR_JSON = "/about/facts_not_found.json";
 
+    private AboutListRobot.Eyes eyes = new AboutListRobot.Eyes();
     private MockWebServer mockWebServer;
 
     @Before
     public void setUp() throws Exception {
         mockWebServer = new MockWebServer();
         mockWebServer.start();
-        APIClient.BASE_URL = mockWebServer.url("/").toString();
+        APIClient.BASE_URL = mockWebServer.url(URL_PATH).toString();
     }
 
     @Test
     public void onLaunch_apiSuccess_seesAboutContent() {
         mockWebServer.enqueue(new MockResponse()
                 .setResponseCode(HTTP_OK)
-                .setBody(FileUtils.getResponseFromFile(getClass(), "/about/facts.json")));
+                .setBody(FileUtils.getResponseFromFile(getClass(), ABOUT_FACTS_CONTENT_JSON)));
 
         activityTestRule.launchActivity(new Intent());
 
@@ -54,7 +57,7 @@ public class AboutCanadaListActivityTest {
     public void onLaunch_apiFailure_seesNoContentInfo() {
         mockWebServer.enqueue(new MockResponse()
                 .setResponseCode(HTTP_UNAVAILABLE)
-                .setBody(FileUtils.getResponseFromFile(getClass(), "/about/facts_not_found.json")));
+                .setBody(FileUtils.getResponseFromFile(getClass(), ABOUT_FACTS_ERROR_JSON)));
 
         activityTestRule.launchActivity(new Intent());
