+++ app/src/androidTest/java/com/hcl/canada/utils/RoboUtils.java
@@ -1,7 +1,6 @@
-package utils;
+package com.hcl.canada.utils;
 
 import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.action.ViewActions.swipeDown;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
@@ -15,10 +14,5 @@ public class RoboUtils {
 
     public static void doesNotSeesViewWithId(int id) {
         onView(withId(id)).check(matches(not(isDisplayed())));
-    }
-
-    public static void swipeDownRefreshLayout(int id) {
-        onView(withId(id))
-                .perform(swipeDown());
     }
 }
