+++ app/src/androidTest/resources/about/facts.json
@@ -15,7 +15,7 @@
       "title": "Transportation",
       "description": "It is a well known fact that polar bears are the main mode of transportation in Canada. They consume far less gas and have the added benefit of being difficult to steal.",
       "imageHref": "http://1.bp.blogspot.com/_VZVOmYVm68Q/SMkzZzkGXKI/AAAAAAAAADQ/U89miaCkcyo/s400/the_golden_compass_still.jpg"
-    },ō
+    },
     {
       "title": "Hockey Night in Canada",
       "description": "These Saturday night CBC broadcasts originally aired on radio in 1931. In 1952 they debuted on television and continue to unite (and divide) the nation each week.",
