+++ app/src/androidTest/resources/about/facts_not_found.json
@@ -3,4 +3,4 @@
     "code": 503,
     "message": "Facts Not found"
   }
-}ō
+}
