+++ app/src/main/java/com/hcl/canada/presenters/ViewStatePresenter.java
@@ -1,4 +1,4 @@
-package com.hcl.canada.ui.presenters;
+package com.hcl.canada.presenters;
 
 import android.support.annotation.NonNull;
